move panel code and dependencies to shared/
