+++ shared/src/commandPalette/CommandList.scss
