+++ shared/src/components/Resizable.scss
