+++ shared/src/components/Resizable.tsx
