+++ shared/src/components/Tabs.scss
