+++ shared/src/components/Toggle.scss
