+++ shared/src/extensions/ExtensionStatus.scss
