--- shared/src/index.css
@@ -1,6 +1,0 @@
-@import url('./commandPalette/CommandList.css');
-@import url('./components/PopoverButton.css');
-@import url('./components/Toggle.css');
-@import url('./extensions/ExtensionStatus.css');
-@import url('./notifications/NotificationItem.css');
-@import url('./notifications/Notifications.css');
