+++ shared/src/index.scss
@@ -1,0 +1,13 @@
+@import './commandPalette/CommandList';
+@import './components/Markdown';
+@import './components/PopoverButton';
+@import './components/Resizable';
+@import './components/Toggle';
+@import './components/Tabs';
+@import './extensions/ExtensionStatus';
+@import './notifications/NotificationItem';
+@import './notifications/Notifications';
+@import './panel/Panel';
+@import './panel/views/FileLocations';
+@import './panel/views/FileLocationsTree';
+/* TODO!(sqs): ensure no CSS class name conflict with github/gitlab for browser ext */
