+++ shared/src/notifications/NotificationItem.scss
