+++ shared/src/notifications/Notifications.scss
