+++ shared/src/panel/Panel.scss
@@ -1,6 +1,4 @@
 @import '../components/Resizable';
-@import '../repo/blob/panel/BlobPanel';
-@import '../repo/blob/panel/FileLocations';
 
 .panel {
     flex: 1 1 50%;
