+++ shared/src/panel/views/FileLocations.scss
