+++ shared/src/panel/views/FileLocations.tsx
@@ -5,14 +5,14 @@ import MapSearchIcon from 'mdi-react/MapSearchIcon'
 import * as React from 'react'
 import { combineLatest, merge, Observable, of, Subject, Subscription } from 'rxjs'
 import { catchError, delay, distinctUntilChanged, map, startWith, switchMap, takeUntil } from 'rxjs/operators'
-import { parseRepoURI } from '../..'
-import { Location } from '../../../../../shared/src/api/protocol/plainTypes'
-import { asError } from '../../../../../shared/src/util/errors'
-import { ErrorLike, isErrorLike } from '../../../../../shared/src/util/errors'
-import { FileMatch, IFileMatch, ILineMatch } from '../../../components/FileMatch'
-import { VirtualList } from '../../../components/VirtualList'
-import { propertyIsDefined } from '../../../util/types'
-import { toPrettyBlobURL, toRepoURL } from '../../../util/url'
+import { FileMatch, IFileMatch, ILineMatch } from '../../../../web/src/components/FileMatch'
+import { VirtualList } from '../../../../web/src/components/VirtualList'
+import { parseRepoURI } from '../../../../web/src/repo'
+import { propertyIsDefined } from '../../../../web/src/util/types'
+import { toPrettyBlobURL, toRepoURL } from '../../../../web/src/util/url'
+import { Location } from '../../api/protocol/plainTypes'
+import { asError } from '../../util/errors'
+import { ErrorLike, isErrorLike } from '../../util/errors'
 
 export const FileLocationsError: React.FunctionComponent<{ pluralNoun: string; error: ErrorLike }> = ({
     pluralNoun,
