+++ shared/src/panel/views/FileLocationsTree.scss
