+++ shared/src/util/linkClickHandler.ts
