+++ web/src/SourcegraphWebApp.scss
@@ -253,7 +253,6 @@ hr {
 @import './components/CodeExcerpt';
 @import './components/DecoratedTextLines';
 @import './components/HeroPage';
-@import './components/Markdown';
 @import './components/ModalPage';
 @import './components/MonacoEditor';
 @import './components/PopoverButton';
