+++ web/src/docSite/DocSitePage.tsx
@@ -6,16 +6,16 @@ import { Link, RouteComponentProps } from 'react-router-dom'
 import { Observable, Subject, Subscription } from 'rxjs'
 import { catchError, distinctUntilChanged, map, startWith, switchMap, tap } from 'rxjs/operators'
 import { LinkOrSpan } from '../../../shared/src/components/LinkOrSpan'
+import { Markdown } from '../../../shared/src/components/Markdown'
 import { gql } from '../../../shared/src/graphql/graphql'
 import * as GQL from '../../../shared/src/graphql/schema'
 import { asError, createAggregateError, ErrorLike, isErrorLike } from '../../../shared/src/util/errors'
+import { createLinkClickHandler } from '../../../shared/src/util/linkClickHandler'
 import { memoizeObservable } from '../../../shared/src/util/memoizeObservable'
 import { queryGraphQL } from '../backend/graphql'
 import { HeroPage } from '../components/HeroPage'
-import { Markdown } from '../components/Markdown'
 import { PageTitle } from '../components/PageTitle'
 import { eventLogger } from '../tracking/eventLogger'
-import { createLinkClickHandler } from '../util/linkClickHandler'
 
 const queryDocPage = memoizeObservable(
     (path: string): Observable<GQL.IDocSitePage | null> =>
