+++ web/src/extensions/explore/ExtensionViewsExploreSection.tsx
@@ -5,9 +5,9 @@ import { Subscription } from 'rxjs'
 import { map } from 'rxjs/operators'
 import { ContributableViewContainer } from '../../../../shared/src/api/protocol'
 import { PanelView } from '../../../../shared/src/api/protocol/plainTypes'
+import { Markdown } from '../../../../shared/src/components/Markdown'
 import { ExtensionsControllerProps } from '../../../../shared/src/extensions/controller'
-import { Markdown } from '../../components/Markdown'
-import { createLinkClickHandler } from '../../util/linkClickHandler'
+import { createLinkClickHandler } from '../../../../shared/src/util/linkClickHandler'
 
 interface Props extends ExtensionsControllerProps {
     history: H.History
