+++ web/src/extensions/shared.scss
@@ -1,4 +1,4 @@
-@import '../../shared/src/index.css';
+@import '../../shared/src/index';
 
 :root {
     --primary-opacity-2: #{transparentize($primary, 0.8)};
