+++ web/src/global/GlobalAlerts.tsx
@@ -1,9 +1,9 @@
 import marked from 'marked'
 import * as React from 'react'
 import { Subscription } from 'rxjs'
+import { Markdown } from '../../../shared/src/components/Markdown'
 import { isSettingsValid, SettingsCascadeProps } from '../../../shared/src/settings/settings'
 import { DismissibleAlert } from '../components/DismissibleAlert'
-import { Markdown } from '../components/Markdown'
 import { Settings } from '../schema/settings.schema'
 import { SiteFlags } from '../site'
 import { siteFlags } from '../site/backend'
