+++ web/src/repo/RepoRevContainer.scss
@@ -1,4 +1,3 @@
-@import '../components/Resizable';
 @import '../tree/Tree';
 @import './blob/BlobPage';
 @import './TreePage';
