+++ web/src/repo/RepoRevSidebar.tsx
@@ -4,10 +4,15 @@ import * as React from 'react'
 import { fromEvent, Subscription } from 'rxjs'
 import { filter } from 'rxjs/operators'
 import { AbsoluteRepoFile } from '.'
+import { Resizable } from '../../../shared/src/components/Resizable'
+import {
+    Spacer,
+    Tab,
+    TabBorderClassName,
+    TabsWithLocalStorageViewStatePersistence,
+} from '../../../shared/src/components/Tabs'
 import { ExtensionsControllerProps } from '../../../shared/src/extensions/controller'
 import * as GQL from '../../../shared/src/graphql/schema'
-import { Resizable } from '../components/Resizable'
-import { Spacer, Tab, TabBorderClassName, TabsWithLocalStorageViewStatePersistence } from '../components/Tabs'
 import { eventLogger } from '../tracking/eventLogger'
 import { Tree } from '../tree/Tree'
 import { FormatListBulletedIcon } from '../util/icons' // TODO: Switch to mdi icon
