+++ web/src/repo/RevisionsPopover.scss
@@ -1,5 +1,4 @@
 @import '../components/ConnectionPopover';
-@import '../components/Tabs';
 
 .revisions-popover {
     width: 35rem;
