+++ web/src/repo/RevisionsPopover.tsx
@@ -4,13 +4,13 @@ import { Link } from 'react-router-dom'
 import { Observable } from 'rxjs'
 import { map } from 'rxjs/operators'
 import { replaceRevisionInURL } from '.'
+import { TabsWithLocalStorageViewStatePersistence } from '../../../shared/src/components/Tabs'
 import { gql } from '../../../shared/src/graphql/graphql'
 import * as GQL from '../../../shared/src/graphql/schema'
 import { createAggregateError } from '../../../shared/src/util/errors'
 import { memoizeObservable } from '../../../shared/src/util/memoizeObservable'
 import { queryGraphQL } from '../backend/graphql'
 import { FilteredConnection, FilteredConnectionQueryArgs } from '../components/FilteredConnection'
-import { TabsWithLocalStorageViewStatePersistence } from '../components/Tabs'
 import { eventLogger } from '../tracking/eventLogger'
 import { CircleChevronLeftIcon } from '../util/icons' // TODO: Switch to mdi icon
 import { GitRefNode, queryGitRefs } from './GitRef'
