+++ web/src/repo/blob/BlobPage.scss
@@ -1,5 +1,4 @@
 @import './Blob';
-@import '../../panel/Panel';
 @import './RenderedFile';
 
 .blob-page {
