+++ web/src/repo/blob/panel/BlobPanel.scss
@@ -1,3 +1,1 @@
-@import './FileLocations';
-@import './FileLocationsTree';
 @import '../discussions/DiscussionsTree.scss';
